test: Adds test for deletion in tree_test
Ensure that the final coverage for nodes with both links (which may also
have child nodes) is in place for tree_test.

Other main thing in this commit is changing the node equal_to? method
to be include? similar to the Tree class. equal_to? was getting confusing
when I was trying to use it in other places.